+++ TMS-frontend/src/component/sidebar/SidebarElements.js
@@ -14,9 +14,8 @@ align-items: center;
 top: 0;
 left: 0;
 transition: 0.3s ease-in-out;
-/* opacity: ${({ isOpen })=>(isOpen ? '100%' : '0')};
-top : ${({ isOpen})=> (isOpen ? '0': '-100%')}; */
-top: 0;
+opacity: ${({ isOpen })=>(isOpen ? '100%' : '0')};
+top : ${({ isOpen})=> (isOpen ? '0': '-100%')};
 
 `;
 export const CloseIcon = styled(FaTimes)`
